Reintroduce layer taps on space and enter keys
Buth this time only for peeking into layer 2. Because it is expected
that layer 2 would rarely require use of both hands simultaneously. For
a permanent switch we have the layer change key anyways.

LAyer 1 is anyways just one key away and might require lot of two handed
operations hence no need for layer-tap for it.