FIX:not able to jump to base from multimedia/system
The `&trans` seems to transfer other keys like backspace into lower
layers without issues but for some reason `&to 0` from symbols layer
is being transferred into lower layers, that is, navmum and system. We
try to find the reason later, for now we explicitly set the key in the
affected layers.